Added speechIntent constants
I added the speechIntents object containing several constants that
extensions can register to handle.  This allows extensions access to our
complex grammars in a very simple way.  If more than two extensions are
registered for the same speechIntent, then the user is given a choice as
to which to use.  That preference is then saved until another app
subscribes to that speechIntent.